+++ src/pyke/ddragon/feats.py
@@ -2,15 +2,15 @@ from typing import Any
 
 from .._base_data_dragon_client import _BaseDataDragonClient
 
-class ChallengesData:
+class FeatsData:
     def __init__(self, client: _BaseDataDragonClient):
         self._client = client
 
     def get_all(self, locale: str) -> dict[str, Any]:
-        """# Get all challenges by locale
+        """# Get all feats by locale
 
         **Example:**
-            `challenges = ddragon.challenges.get_all("en_GB")`
+            `feats = ddragon.feats.get_all("en_GB")`
 
         **Args:**
             `locale (str)` Locale to use.
@@ -19,4 +19,4 @@ class ChallengesData:
             `dict[str, any]`
         """  # fmt: skip
 
-        return self._client._data_dragon_cdn_request(locale, "challenges")
+        return self._client._data_dragon_cdn_request(locale, "feats")